+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -3,11 +3,10 @@ package woowacourse.shopping.ui.cart
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import woowacourse.shopping.di.Inject
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.repository.CartRepository
 
-class CartViewModel @Inject constructor(
+class CartViewModel constructor(
     private val cartRepository: CartRepository,
 ) : ViewModel() {
